+++ d60.Cirqus.PostgreSql/PostgreSqlEventStore.cs
@@ -73,7 +73,7 @@ END$$;
             }
         }
 
-        public void Save(Guid batchId, IEnumerable<Event> batch)
+        public void Save(Guid batchId, IEnumerable<EventData> batch)
         {
             var eventList = batch.ToList();
 
@@ -168,7 +168,7 @@ INSERT INTO ""{0}"" (
             return connection;
         }
 
-        public IEnumerable<Event> Load(Guid aggregateRootId, long firstSeq = 0)
+        public IEnumerable<EventData> Load(Guid aggregateRootId, long firstSeq = 0)
         {
             using (var connection = GetConnection())
             {
@@ -196,7 +196,7 @@ INSERT INTO ""{0}"" (
             }
         }
 
-        public IEnumerable<Event> Stream(long globalSequenceNumber = 0)
+        public IEnumerable<EventData> Stream(long globalSequenceNumber = 0)
         {
             using (var connection = GetConnection())
             {
@@ -223,12 +223,12 @@ SELECT ""data"", ""meta"" FROM ""{0}"" WHERE ""globSeqNo"" >= @cutoff ORDER BY "
             }
         }
 
-        Event ReadEvent(IDataRecord reader)
+        EventData ReadEvent(IDataRecord reader)
         {
             var data = (byte[]) reader["data"];
             var meta = (byte[]) reader["meta"];
 
-            return Event.FromMetadata(_metadataSerializer.Deserialize(Encoding.UTF8.GetString(meta)), data);
+            return EventData.FromMetadata(_metadataSerializer.Deserialize(Encoding.UTF8.GetString(meta)), data);
         }
 
         public long GetNextGlobalSequenceNumber()