+++ d60.Cirqus.SQLite/SQLiteEventStore.cs
@@ -64,7 +64,7 @@ namespace d60.Cirqus.SQLite
             public byte[] Meta { get; set; }
         }
 
-        public void Save(Guid batchId, IEnumerable<Events.Event> batch)
+        public void Save(Guid batchId, IEnumerable<Events.EventData> batch)
         {
             var nextGlobalSequenceNumber = GetNextGlobalSequenceNumber();
 
@@ -117,24 +117,24 @@ namespace d60.Cirqus.SQLite
             }
         }
 
-        public IEnumerable<Events.Event> Load(Guid aggregateRootId, long firstSeq = 0)
+        public IEnumerable<Events.EventData> Load(Guid aggregateRootId, long firstSeq = 0)
         {
             // must be foreach here - SQLite's LINQ thingie does not play well with _domainEventSerializer.Deserialize(e.EventData)
             foreach (var e in _connection.Table<Event>()
                 .Where(e => e.AggregateRootId == aggregateRootId)
                 .Where(e => e.SequenceNumber >= firstSeq))
             {
-                yield return Events.Event.FromMetadata(_metadataSerializer.Deserialize(Encoding.UTF8.GetString(e.Meta)), e.Data);
+                yield return Events.EventData.FromMetadata(_metadataSerializer.Deserialize(Encoding.UTF8.GetString(e.Meta)), e.Data);
             }
         }
 
-        public IEnumerable<Events.Event> Stream(long globalSequenceNumber = 0)
+        public IEnumerable<Events.EventData> Stream(long globalSequenceNumber = 0)
         {
             // must be foreach here - SQLite's LINQ thingie does not play well with _domainEventSerializer.Deserialize(e.EventData)
             foreach (var e in _connection.Table<Event>()
                 .Where(e => e.GlobalSequenceNumber >= globalSequenceNumber))
             {
-                yield return Events.Event.FromMetadata(_metadataSerializer.Deserialize(Encoding.UTF8.GetString(e.Meta)), e.Data);
+                yield return Events.EventData.FromMetadata(_metadataSerializer.Deserialize(Encoding.UTF8.GetString(e.Meta)), e.Data);
             }
         }