+++ d60.Cirqus.Tests/Events/Ntfs/TestNtfsEventStore.cs
@@ -27,7 +27,7 @@ namespace d60.Cirqus.Tests.Events.Ntfs
             // make one full commit
             _eventStore.Save(rootId, new[]
             {
-                Event.FromMetadata(new Metadata
+                EventData.FromMetadata(new Metadata
                 {
                     {DomainEvent.MetadataKeys.SequenceNumber, 0.ToString(Metadata.NumberCulture)},
                     {DomainEvent.MetadataKeys.AggregateRootId, rootId.ToString()}
@@ -37,7 +37,7 @@ namespace d60.Cirqus.Tests.Events.Ntfs
 
             // save an event to a file, without committing
             _eventStore.DataStore.Write(
-                Event.FromMetadata(new Metadata
+                EventData.FromMetadata(new Metadata
                 {
                     {DomainEvent.MetadataKeys.SequenceNumber, 1.ToString(Metadata.NumberCulture)},
                     {DomainEvent.MetadataKeys.AggregateRootId, rootId.ToString()},
@@ -57,7 +57,7 @@ namespace d60.Cirqus.Tests.Events.Ntfs
             // make one full commit
             _eventStore.Save(rootId, new[]
             {
-                Event.FromMetadata(new Metadata
+                EventData.FromMetadata(new Metadata
                 {
                     {DomainEvent.MetadataKeys.SequenceNumber, 0.ToString(Metadata.NumberCulture)},
                     {DomainEvent.MetadataKeys.AggregateRootId, rootId.ToString()}
@@ -87,7 +87,7 @@ namespace d60.Cirqus.Tests.Events.Ntfs
             // make one full commit
             _eventStore.Save(rootId, new[]
             {
-                Event.FromMetadata(new Metadata
+                EventData.FromMetadata(new Metadata
                 {
                     {DomainEvent.MetadataKeys.SequenceNumber, 0.ToString(Metadata.NumberCulture)},
                     {DomainEvent.MetadataKeys.AggregateRootId, rootId.ToString()}
@@ -108,7 +108,7 @@ namespace d60.Cirqus.Tests.Events.Ntfs
             // make one full commit
             _eventStore.Save(rootId, new[]
             {
-                Event.FromMetadata(new Metadata
+                EventData.FromMetadata(new Metadata
                 {
                     {DomainEvent.MetadataKeys.SequenceNumber, 1.ToString(Metadata.NumberCulture)},
                     {DomainEvent.MetadataKeys.AggregateRootId, rootId.ToString()}
@@ -131,7 +131,7 @@ namespace d60.Cirqus.Tests.Events.Ntfs
             // make one full commit
             _eventStore.Save(rootId, new[]
             {
-                Event.FromMetadata(new Metadata
+                EventData.FromMetadata(new Metadata
                 {
                     {DomainEvent.MetadataKeys.SequenceNumber, 0.ToString(Metadata.NumberCulture)},
                     {DomainEvent.MetadataKeys.AggregateRootId, rootId.ToString()}
@@ -139,7 +139,7 @@ namespace d60.Cirqus.Tests.Events.Ntfs
             });
 
             // make one that fails right after index write
-            var domainEvent = Event.FromMetadata(new Metadata
+            var domainEvent = EventData.FromMetadata(new Metadata
             {
                 {DomainEvent.MetadataKeys.SequenceNumber, 1.ToString(Metadata.NumberCulture)},
                 {DomainEvent.MetadataKeys.AggregateRootId, rootId.ToString()},
@@ -152,7 +152,7 @@ namespace d60.Cirqus.Tests.Events.Ntfs
             // make one full commit
             _eventStore.Save(rootId, new[]
             {
-                Event.FromMetadata(new Metadata
+                EventData.FromMetadata(new Metadata
                 {
                     {DomainEvent.MetadataKeys.SequenceNumber, 1.ToString(Metadata.NumberCulture)},
                     {DomainEvent.MetadataKeys.AggregateRootId, rootId.ToString()}