+++ d60.Cirqus.Tests/Snapshotting/TestSnapshottingWithFairlyLongHistory.cs
@@ -150,7 +150,7 @@ caching in use: {3}",
                 return InnerAggregateRootRepository.Exists<TAggregate>(aggregateRootId, maxGlobalSequenceNumber, unitOfWork);
             }
 
-            public IEnumerable<Event> Load(Guid aggregateRootId, long firstSeq = 0)
+            public IEnumerable<EventData> Load(Guid aggregateRootId, long firstSeq = 0)
             {
                 var stopwatch = Stopwatch.StartNew();
 
@@ -161,7 +161,7 @@ caching in use: {3}",
                 return domainEvents;
             }
 
-            public IEnumerable<Event> Stream(long globalSequenceNumber = 0)
+            public IEnumerable<EventData> Stream(long globalSequenceNumber = 0)
             {
                 return InnerEventStore.Stream(globalSequenceNumber);
             }
@@ -171,7 +171,7 @@ caching in use: {3}",
                 return InnerEventStore.GetNextGlobalSequenceNumber();
             }
 
-            public void Save(Guid batchId, IEnumerable<Event> events)
+            public void Save(Guid batchId, IEnumerable<EventData> events)
             {
                 var stopwatch = Stopwatch.StartNew();