+++ d60.Cirqus/Diagnostics/DiagnosticsConfigurationExtensions.cs
@@ -158,12 +158,12 @@ namespace d60.Cirqus.Diagnostics
                 _operationProfiler = operationProfiler;
             }
 
-            public IEnumerable<Event> Load(Guid aggregateRootId, long firstSeq = 0)
+            public IEnumerable<EventData> Load(Guid aggregateRootId, long firstSeq = 0)
             {
                 return _innerEventStore.Load(aggregateRootId, firstSeq);
             }
 
-            public IEnumerable<Event> Stream(long globalSequenceNumber = 0)
+            public IEnumerable<EventData> Stream(long globalSequenceNumber = 0)
             {
                 return _innerEventStore.Stream(globalSequenceNumber);
             }
@@ -181,7 +181,7 @@ namespace d60.Cirqus.Diagnostics
                 }
             }
 
-            public void Save(Guid batchId, IEnumerable<Event> events)
+            public void Save(Guid batchId, IEnumerable<EventData> events)
             {
                 var stopwatch = Stopwatch.StartNew();
                 try